TONY: Bugfix for OFFSETOF macro on 64-bit systems

+++ engines/tony/mpal/memory.cpp
@@ -64,7 +64,7 @@ void *MemoryManager::alloc(uint32 size, uint flags) {
 	return &item->_data[0];
 }
 
-#define OFFSETOF(type, field)    ((unsigned long) &(((type *) 0)->field))
+#define OFFSETOF(type, field)    ((size_t) &(((type *) 0)->field))
 
 /**
  * Returns a reference to the MemoryItem for a gien byte pointer
